Add feature testing framework
Use python's `behave` to test the commandline client from the outside.

The only actual tests so far are for the ability to quit the application
and that the application should wait for the user input.
Test-drove the actual minimal C# code to meet the requirements.
Hooked up the test launchers to both Travis and AppVeyor for test on
Linux (via Mono) and Windows (MS .Net),
Described the requirements for launching the tests in the README file.
Added some more background info on why `behave` was chosen there as
well.