Add ability to read posted messages
Anyone can view particular user's messages by typing in their user name
now. The messages are sorted in reverse chronological order and
displayed with rough notion of relative time that has passes since the
message has been written.

Coding-wise - all of this is a particularly big change, as this was a first 'real' feature to
implement. There is most of the internal infrastructure being added in
this commit.

General architecture goes as follows:

There is MainLoop. Per each line of the user input it asks
CommandExtractor to create a new command corresponding to that line.
It then executes the command and outputs any result to the output
stream.

Messages posted are now stored in-memory, in TimelineRepository.
PostingCommand adds messages there while ReadingCommand reads from
there.